fix: исправил хуки плеера, сделал отдельный хук добавления в избранное, разделил тесты
