--- src/hooks/use-player-volume.ts
@@ -1,13 +1,0 @@
-import { useEffect } from "react";
-
-export const usePlayerVolume = (
-  audioRef: React.RefObject<HTMLAudioElement | null>,
-  volume: number,
-) => {
-  useEffect(() => {
-    const audio = audioRef.current;
-    if (audio) {
-      audio.volume = volume;
-    }
-  }, [audioRef, volume]);
-};
